Modify user filtering algorithm.
Filtering users by role on the database has the potential to make for a very expensive query, as it performs a LIKE on an unindexed field.

But if the users were all brought over at once, unfiltered, the potential exists for an array with millions of elements, straining the application server's memory limit.

So the compromise is: batching.

Users are fetched from the database, unfiltered, in batches of user-defined size. They are then filtered, and a new batch is brought in.

This algorithm looks to strike a balance between load on the database server, and memory limits on the application server.