Add --not-a-dry-run flag controlling whether or not the command performs data modifications.
By default, the command outputs the IDs of the users it would have attempted to generate guest authors for, had this not been a dry run.

Passing in the --not-a-dry-run flag tells the code to actually make the attempts.

Props sboisvert for the idea to make this command dry by default.