Add wp-cli command to delete guest authors, with option to reassign their posts.
Currently, there is no way to delete guest authors from the command line. This command fills that need.

It is activated using `delete-guest-authors`, and passing in the path to a CSV file containing a list of two columns: the ID to delete, and the login to assign the posts to (or "false" to not reassign).

The command runs dry by default, showing the end user the guest authors it would have attempted to delete, and the user login it would have attempted to transfer the posts to.

Passing --no-dry-run, and confirming that you really do want to delete guest authors, actively modifies the database.