regenerate pot file
Allow for remapping to user named "false".

The CAP delete function takes, as its second parameter, either a string containing the login of the user you wish to reassign the deleted guest author's posts to, or a boolean false stating that you do wish to reassign.

But what if you want to reassign to a user named false? Before that was not possible.

Now, the code specifically looks for if the instance has a user named false, and if so, allows it to be reassigned to. If not, it assumes the boolean was intended.